requirements.txt
fastapi
uvicorn[standard]
ultralytics
python-multipart
pillow
torch
